+++ src/app/src/constants.js
@@ -1,2 +1,3 @@
 export const MAP_CENTER = [35.7596, -79.0193]; // North Carolina
 export const MAP_INITIAL_ZOOM = 13;
+export const INITIAL_POLYGON_SCALE_FACTOR = 0.5;